Fix: state정리 및 나의 별점 추가 - 상위 컴포넌트로 state 이동 - 나의 별점 UI 추가
