+++ src/pages/Contents/InfoAct/Stars/RightStar.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
-import { ReactComponent as Star } from '../../../assets/icons/star4.svg';
+import { ReactComponent as Star } from '../../../../assets/icons/star4.svg';
+import './RightStar.scss';
 
 class RatingStars extends Component {
   render() {
@@ -7,6 +8,7 @@ class RatingStars extends Component {
 
     return (
       <Star
+        className="right-star"
         onClick={() => onClick(index)}
         onMouseEnter={() => onMouseEnter(index)}
         onMouseLeave={onMouseLeave}
