+++ src/pages/Contents/InfoAct/Stars/RightStar.scss
