+++ src/pages/Contents/MovieInfo/MovieInfo.js
@@ -4,6 +4,16 @@ import './MovieInfo.scss';
 
 class MovieInfo extends Component {
   render() {
+    const {
+      onClick,
+      onClickBtn,
+      onMouseEnter,
+      onMouseLeave,
+      ratingComment,
+      setRating,
+      setHoverRating,
+    } = this.props;
+
     return (
       <section className="movie-info">
         <div className="movie-poster-all">
@@ -18,8 +28,21 @@ class MovieInfo extends Component {
             <h1 className="title">샹치와 텐 링즈의 전설</h1>
             <p className="text">2021 ・ 액션/모험/판타지 ・ 미국/중국/호주</p>
           </div>
-          <span className="rating-star">평균 ★4.2(988명)</span>
-          <InfoAct />
+          <span className="rating-star">
+            평균 ★4.2(988명)
+            <span className="rating-mystar">
+              {setRating !== 0 ? `나의 평점 ★${setRating}` : ''}
+            </span>
+          </span>
+          <InfoAct
+            onClick={onClick}
+            onClickBtn={onClickBtn}
+            onMouseEnter={onMouseEnter}
+            onMouseLeave={onMouseLeave}
+            ratingComment={ratingComment}
+            setRating={setRating}
+            setHoverRating={setHoverRating}
+          />
         </article>
       </section>
     );
