+++ src/pages/Contents/MovieInfo/MovieInfo.scss
@@ -55,6 +55,11 @@
         margin-top: 8px;
         border-bottom: 1px solid #f0f0f0;
       }
+
+      .rating-mystar {
+        color: #ff2f6e;
+        margin-left: 16px;
+      }
     }
   }
 }
